requirements.txt
# Core pipeline
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0

# API (FastAPI + Zapier MCP)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
